+++ frontend/src/views/ElementSearchResultsPage/ResultCard.vue
@@ -2,12 +2,12 @@
   <div class="fr-card fr-enlarge-link">
     <div class="fr-card__body">
       <div class="fr-card__content">
-        <div class="fr-card__title" style="order: unset">
+        <div class="fr-card__title order-none">
           <router-link :to="route" class="fr-card__link capitalize">{{ result.name }}</router-link>
         </div>
-        <div class="mt-2 flex">
-          <div><v-icon scale="0.85" class="mr-1" :name="icon" /></div>
-          <div style="margin-top: 1px">{{ type }}</div>
+        <div class="mt-2 flex gap-x-1">
+          <div><v-icon scale="0.85" :name="icon" /></div>
+          <div class="mt-[1px]">{{ type }}</div>
         </div>
       </div>
     </div>
