+++ frontend/src/views/ElementSearchResultsPage/index.vue
